exporter/app/requirements.txt
pyserial~=3.5
prometheus_client~=0.22.1